requirements.txt
accelerate==0.33.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asciitree==0.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
dask==2024.6.0
decord==0.6.0
diffusers==0.30.0
docker-pycreds==0.4.0
einops==0.7.0
fasteners==0.19
filelock==3.15.3
fsspec==2024.6.0
ftfy==6.2.3
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.23.4
hydra-core==1.3.2
idna==3.7
imageio==2.21.2
imageio_ffmpeg==0.5.1
importlib_metadata==7.2.0
Jinja2==3.1.4
locket==1.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
moviepy==2.1.1
mpmath==1.3.0
networkx==3.3
numcodecs==0.12.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
packaging==24.1
pandas==2.2.1
partd==1.4.2
pillow==10.3.0
protobuf==4.25.3
psutil==6.0.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
robosuite==1.4.1
safetensors==0.4.4
scipy==1.14.0
sentry-sdk==2.6.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.12.1
timm==1.0.9
tokenizers==0.19.1
toolz==0.12.1
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.66.2
transformers==4.44.0
triton==2.2.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wandb==0.19.1
wcwidth==0.2.13
zarr==2.18.2
zipp==3.19.2
